# Optimize collision generation performance