… test_large_curvy.tscn