The time needed to generate collision shapes depends on two factors: edge generation and the convex decomposition of the polygon by `CollisionPolygon2D`.
Especially the latter is very expensive on large shapes that are difficult to decompose.
Additionally the amount of points the collision polygon consists of depends on the tesselation settings.

This patch tweaks the default tesselation settings to provide pretty much the same visual quality at less computation time.
I think they can be even lowered more to `tesselation_tolerance = 7.0` and maybe `tesselation_stages = 3`. 
I think it's better to provide faster defaults and let users increase them if necessary. Need feedback on this.
This also speeds up regular edge generation respectively.

Edit: Settled on `tesselation_stages = 3` but kept `tesselation_tolerance = 6.0`.

The PR also adds two new properties to `shape.gd`: `collision_generation_method` and `collision_update_mode`.
See also property doc comments.
I set the defaults to `Fast` (curve based generation instead of edge based) and `Editor` (only update in editor), as these settings probably fit most projects and improve performance and load times immensely.

To mention some numbers, consider the included `test_large_curvy.tscn` scene.

| Setting                     | Before           | After  |
|-----------------------------|------------------|--------|
| tesselation_tolerance       | 4.0              | 6.0    |
| tesselation_stages          | 5                | 4      |
| collision_generation_method | Precise          | Fast   |
| collision_update_mode       | EditorAndRuntime | Editor |


| Metric           | Before | After    |
|------------------|-------------|----|
| Num points       | 1366   | 918
| Generate         | 38 ms  | 0.007 ms
| Convex decompose | 64 ms  | 32 ms

With the new settings, this scene updates around 60 ms faster in editor and loads 102 ms faster during runtime (disregarding improved edge generation time through tesselation settings).
Since generation time became negligible with `Fast` collision generation and convex decomposition is not in our hands, this is as fast as collision generation will get.
For even more performance, users can tweak tesselation settings according to their project needs.


With this patch merged, I'd also close #119, as the load times are absolutely fine by now.