@mphe 2024-06-19
> I think they can be even lowered more to tesselation_tolerance = 7.0 and maybe tesselation_stages = 3.

What do you think about lowering the settings more?