@limbonaut 2024-06-19
> What do you think about lowering the settings more?

In my tests, increasing tolerance to 7.0 produced some drastic changes on some large segments. Personally, I wouldn't like it as a default if it's prone to such errors.

![image](https://github.com/SirRamEsq/SmartShape2D/assets/2766569/1c483273-ee2b-4f4c-b8fc-f01a20bca46b)
↓↓↓↓↓↓↓    TURNS FLAT  ↓↓↓↓↓↓↓
![image](https://github.com/SirRamEsq/SmartShape2D/assets/2766569/69e059fd-2743-44cb-a360-9202e940ae65)

Maybe we could turn it into a project setting?