@mphe 2024-06-19
What about `tesselation_stages = 3`?

Project setting sounds good, but I'd also keep the properties for overriding the project settings when needed, e.g. when having a shape that requires more detail.
This will be a breaking change though, as the existing properties should be renamed so not all existing shapes automatically override the project settings.
But I think people rarely touched the setting in the first place.