@limbonaut 2024-06-19
> What about `tesselation_stages = 3`?

Feels like a good compromise. In some cases, even looks better for whatever reason.
 
> Project setting sounds good, but I'd also keep the properties for overriding the project settings when needed, e.g. when having a shape that requires more detail. This will be a breaking change though, as the existing properties should be renamed so not all existing shapes automatically override the project settings. But I think people rarely touched the setting in the first place.

Hmm, it could be done with [_get_property_revert](https://docs.godotengine.org/en/stable/classes/class_object.html#class-object-private-method-property-get-revert). I wrote a similar thing for [BlackboardPlan](https://github.com/limbonaut/limboai/blob/8fa609ef9eae0e1cb6351accffb515ac2a9f8dc4/blackboard/blackboard_plan.cpp#L163C1-L180C2) in limboai.
Not sure if it's worth the effort, if people rarely touch these settings.