@limbonaut 2024-06-19
There is probably no reason to have different defaults on different project, is there? The more I think about it, the less I like this idea.