@mphe 2024-06-19
I can think of users having lots of shapes in their game but want to tweak tesselation options globally without having to go through all shapes manually.
But I'm also not sure whether it's actually worth it. Personally, I have my shapes in a separate scene so I just need to tweak the scene to update all shapes. Not sure how other people do it.