@mphe 2024-06-19
I changed the default `tesselation_stages` to `3`. Also fixed a unit test which now requires a slightly greater epsilon.

I think we can hold the idea with the project settings in mind and maybe implement it in the future when there is a demand for it or when there are more settings that would benefit from being project-wide configurable.

In that regard, this PR should be finished.