@mphe 2024-06-20
What do you mean besides the scene? Remove the test scene?